requirements.txt
Django==4.2
djangorestframework==3.14.0
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
# Optional for local Selenium tests (not required for MVP)
selenium==4.10.0
webdriver-manager==3.8.5
